refactor(forge): split GitHub issue methods into github_issues.go (#1025)
Pure same-package code motion: move ListIssues, GetIssue, CreateIssue,
UpdateIssueDescription, UserExists, ListIssueComments, CreateIssueNote,
toGitHubIssue, githubIssueState, and githubIssueStateParam verbatim out of
github.go into the new github_issues.go. No behaviour, signature, body, or
doc-comment change.
